% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{AnimalDataSetProcess}
\alias{AnimalDataSetProcess}
\title{Animal Dataset Process}
\usage{
AnimalDataSetProcess(
  main.dataset,
  group.selected,
  keep.raw = FALSE,
  determine.obs = NA,
  multiple.sample = F
)
}
\arguments{
\item{main.dataset}{Dataset whit multiple files read. The dataset must have the group.factor column, group column,
signal, file and n.obs to be processed}

\item{group.selected}{Group to filter dataset}

\item{keep.raw}{Maintain the raw dataset with all the original data}

\item{determine.obs}{Minimal númber of observations}

\item{multiple.sample}{Use binning to repeat sample from register}
}
\value{
Lis with one or multiple dataset processed
}
\description{
This function process data applying fourirer transformation, filter sample and in case of necesity apply binning to
the signal. it requires a dataset with multiple signals.
Uses file column, hich have filename, to process the different signals.
}
\examples{
\dontrun{AnimalDataSetProcess(archvodeexceldata, 'neotomodon', keep.raw = FALSE)}
}
