% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{AnimalTagTestTraining}
\alias{AnimalTagTestTraining}
\title{Animal Tag Test training}
\usage{
AnimalTagTestTraining(
  dataset.input,
  factor.column = "group.factor",
  multiple.sample = F
)
}
\arguments{
\item{dataset.input}{Input with processed dataset which will be partitionated in training and test groups}

\item{factor.column}{Dataset Column to use it for partitioning}

\item{multiple.sample}{Allos to process segmented files}
}
\value{

}
\description{
This function allows you to tag randomly the test and training groups need for algorithm training phase
}
\examples{
\dontrun{AnimalTagTestTraining(archvodeexceldata, factor.column = "group.factor", multiple.sample = FALSE)}
}
