% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalProcessing.R
\name{BinningSignal}
\alias{BinningSignal}
\title{BinningSignal Signal segmentation}
\usage{
BinningSignal(signal.vector, max.lecture, size.bin)
}
\arguments{
\item{signal.vector}{Vector with the fourier trasnform signal to segmentate}

\item{max.lecture}{Maximum value of the signal}

\item{size.bin}{Size of he bin to be segmented}
}
\value{
Signal with splitted values
}
\description{
Segmentation of signal by bin. It allows to split and group same segments of the signals to compare segments
}
\examples{
\dontrun{BinningSignal(signal.vector, max.lecture = 90, size.bin = 2)}
}
