% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalProcessing.R
\name{calcFFT}
\alias{calcFFT}
\title{Calculate Fourier Transformation}
\usage{
calcFFT(signal.vector, normalize = TRUE, only.ps = FALSE)
}
\arguments{
\item{signal.vector}{Vector with signal data}

\item{normalize}{Allows to normalize the coefficients}

\item{only.ps}{Define if the function returns only the seed power or a dataset with poer and frecuencies}
}
\value{
A seed power vector or a frequency/power dataframe
}
\description{
Get the fourier transformation for a signal vector. The functions allows to normalize the signal dividing by
the number of values. It can return only the seed power or also the position vector
}
\examples{
\dontrun{calcFFT(signal.vector, normalize = TRUE, only.ps = FALSE)}
}
