% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{CleanNames}
\alias{CleanNames}
\title{Clean excel names}
\usage{
CleanNames(string.name.file)
}
\arguments{
\item{string.name.file}{file name to clean}
}
\value{
string with the name of the file without special characters
}
\description{
Allows to clean not valid characters from excel files in order to
denominate each readed dataset. In most of the cases the files have
the patient name of the dataset.
}
\examples{
\dontrun{CleanNames('archivo de excel.xlsx')}
}
