% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{DefineFilteType}
\alias{DefineFilteType}
\title{Find File Type by File Structure}
\usage{
DefineFilteType(file.processed)
}
\arguments{
\item{file.processed}{file head with at least 10 lines of content}
}
\value{
list with the reading file requirements
}
\description{
Some of the files where the animal data is adquired could have different structure.
In order to proces this files. This function allow us to identify the different
structures in order to process each file diferently.
It needs the first 10 lines of the file in order to determine the appropriate
parameters.
}
\examples{
\dontrun{DefineFilteType(archvodeexcel)}
}
