% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{ReadExcelData}
\alias{ReadExcelData}
\title{Read and clean Excel data}
\usage{
ReadExcelData(path, size.of.register = 3e+05)
}
\arguments{
\item{path}{Path to the excel file}

\item{size.of.register}{Maximum number of rows to read from file}
}
\value{
DataFrame/Tibble with file information.
}
\description{
Read excel file with size defined by size.of.register argument.
Cleaned names from the excel filename. And remove NA values.
The File needs to be in specific format, one column with the data.
The default size fo register is 300000 rows, it could be larger but
is the default size of the capture.
}
\examples{
\dontrun{ReadExcelData("~/archivo de excel.xlsx", size.of.register = 300000)}
}
